Refactor structure
And introduce a cable.go file containing the Setup code for the telegram and slack integration
